refactor(newsletter-sign-up): make the newsletter sign up component use the generic step switch component
