+++ components/Main/NewsletterSignUp/SignUpForm/index.tsx
@@ -1,14 +1,13 @@
 import Button from '../../Button'
 import Form from '../../form/Section/StepSwitch/Form'
 import InputField from '../../form/Section/StepSwitch/Form/InputField'
-import { useState } from 'react'
+import { NextContext } from '../../form/Section/StepSwitch'
+import { useContext, useState } from 'react'
 import type { FormEvent } from 'react'
+import type { Next } from '../../form/Section/StepSwitch/types'
 
-export type SignUpFormProps = {
-  next: () => void
-}
-
-export default function SignUpForm({ next }: SignUpFormProps): JSX.Element {
+export default function SignUpForm(): JSX.Element {
+  const next = useContext<Next>(NextContext)
   const [email, setEmail] = useState('')
 
   const handleSubmit = (event: FormEvent<HTMLFormElement>) => {
