+++ components/Main/form/Section/StepSwitch/Confirmation/index.stories.tsx
@@ -6,6 +6,14 @@ const meta: Meta = {
   title: 'components/Main/form/Section/StepSwitch/Confirmation',
 }
 
-export const Default = (): JSX.Element => <Confirmation message="Successfully submitted!">Click Me</Confirmation>
+function mockNext() {
+  console.log('mock next')
+}
+
+export const Default = (): JSX.Element => (
+  <Confirmation message="Successfully submitted!" next={mockNext}>
+    Click Me
+  </Confirmation>
+)
 
 export default meta
