+++ components/Main/form/Section/StepSwitch/Confirmation/index.tsx
@@ -1,11 +1,12 @@
 import Button from '../../../../Button'
 import style from './Confirmation.module.css'
 import { ReactNode } from 'react'
+import type { Next } from '../types'
 
 export type ConfirmationProps = {
   children: ReactNode
   message: string
-  next?: () => void
+  next: Next
 }
 
 export default function Confirmation({ children, message, next }: ConfirmationProps): JSX.Element {
