+++ components/Main/form/Section/StepSwitch/Form/InputField/index.tsx
@@ -1,16 +1,6 @@
 import style from './InputField.module.css'
 import type { ChangeEvent } from 'react'
-
-export type InputFieldProps = {
-  id?: string
-  label?: string
-  name: string
-  placeholder?: string
-  required?: boolean
-  setValue: (value: string) => void
-  type: 'email' | 'password' | 'text'
-  value: string
-}
+import type { InputFieldProps } from './type'
 
 export default function InputField({
   id,
