+++ components/Main/form/Section/StepSwitch/Form/InputField/type.ts
@@ -1,0 +1,10 @@
+export type InputFieldProps = {
+  id?: string
+  label?: string
+  name: string
+  placeholder?: string
+  required?: boolean
+  setValue: (value: string) => void
+  type: 'email' | 'password' | 'text'
+  value: string
+}
