+++ components/Main/form/Section/StepSwitch/Form/TextareaField/types.ts
@@ -1,0 +1,10 @@
+export type TextareaFieldProps = {
+  id: string
+  label?: string
+  maxLength?: number
+  name: string
+  placeholder?: string
+  required?: boolean
+  setValue: (value: string) => void
+  value: string
+}
