+++ components/Main/form/Section/StepSwitch/Form/index.tsx
@@ -3,7 +3,7 @@ import type { FormEventHandler, ReactNode } from 'react'
 
 export type FormProps = {
   children: ReactNode
-  onSubmit?: FormEventHandler
+  onSubmit: FormEventHandler
 }
 
 export default function Form({ children, onSubmit }: FormProps): JSX.Element {
