+++ components/Main/form/Section/StepSwitch/index.tsx
@@ -1,22 +1,14 @@
+import mod from './utilities/mod'
 import { createContext, useState } from 'react'
-import type { ReactNode } from 'react'
-
-export type Next = () => void
-export type StepSwitchProps = {
-  children: Array<ReactNode>
-}
+import type { Next, StepSwitchProps } from './types'
 
 export const NextContext = createContext<Next>(() => {
-  console.log('next step')
+  return undefined
 })
 
-function mod(n: number, m: number) {
-  return ((n % m) + m) % m
-}
-
 export default function StepSwitch({ children }: StepSwitchProps): JSX.Element {
-  const stepLength = children.length
   const [step, setStep] = useState(0)
+  const stepLength = children.length
 
   const next = () => {
     setStep(mod(step + 1, stepLength))
