+++ components/Main/form/Section/StepSwitch/types.ts
@@ -1,0 +1,6 @@
+import type { ReactNode } from 'react'
+
+export type Next = () => void
+export type StepSwitchProps = {
+  children: ReactNode[]
+}
