+++ components/Main/form/Section/StepSwitch/utilities/mod/index.ts
@@ -1,0 +1,5 @@
+function mod(n: number, m: number): number {
+  return ((n % m) + m) % m
+}
+
+export default mod
